Fixing duplicates

+++ main.py
@@ -175,7 +175,8 @@ def load_ignore_list():
     if os.path.exists("ignore_list.txt"):
         with open("ignore_list.txt", "r") as f:
             for line in f:
-                ignore_list.add(line.strip())
+                line = line.replace("\n", "")
+                ignore_list.add(line)
     return ignore_list
 
 
@@ -184,7 +185,8 @@ def load_ignore_macros():
     if os.path.exists("ignore_macros.txt"):
         with open("ignore_macros.txt", "r") as f:
             for line in f:
-                ignore_list.add(line.strip())
+                line = line.replace("\n", "")
+                ignore_list.add(line)
     return ignore_list
 
 
@@ -193,7 +195,8 @@ def load_auto_delete():
     if os.path.exists("auto_delete.txt"):
         with open("auto_delete.txt", "r") as f:
             for line in f:
-                ignore_list.add(line.strip())
+                line = line.replace("\n", "")
+                ignore_list.add(line)
     return ignore_list
 
 
@@ -227,25 +230,34 @@ def cleanup_duplicate_files(comparison_matrix):
         duplicate_files_grouped[short_hash] += [[file_path, file_size, time_create, time_modify]]
     for short_hash in duplicate_files_grouped:
         files = duplicate_files_grouped[short_hash]
-        for file in files:
-            for macro in auto_delete:
-                if macro in file[0]:
-                    for file2 in files:
-                        print(file2)
-                    print(f"Removing {file[0]} because of rule: '{macro}'")
-                    os.remove(file[0])
+        for file in files[:]:
+            if "listro 004" in file[0]:
+                print()
+            if file[0] in ignore_list:
+                try:
                     duplicate_files_grouped[short_hash].remove(file)
-            for ignore_macro in ignore_macros:
-                if ignore_macro in file[0]:
-                    print(f"Ignoring {file[0]}")
-                    ignore_list.add(file[0])
-                    duplicate_files_grouped[short_hash].remove(file)
+                except ValueError:
+                    pass
+            else:
+                for macro in auto_delete:
+                    if macro in file[0]:
+                        for file2 in files:
+                            print(file2)
+                        print(f"Removing {file[0]} because of rule: '{macro}'")
+                        os.remove(file[0])
+                        duplicate_files_grouped[short_hash].remove(file)
+                for ignore_macro in ignore_macros:
+                    if ignore_macro in file[0]:
+                        print(f"Ignoring {file[0]}")
+                        ignore_list.add(file[0])
+
+
 
     write_ignore_list(ignore_list)
 
     for short_hash in duplicate_files_grouped:
         files = duplicate_files_grouped[short_hash]
-        if len(files) == 1:
+        if len(files) < 2:
             continue
         print(f"Opening {len(files)}")
         for file in files:
